requirements.txt
pygame
math
random
pytest
pytorch